[Refactor] #1008 Separated display-scores.cpp/h from scores.cpp/h
