+++ Hengband/Hengband/Hengband.vcxproj
@@ -550,6 +550,7 @@
     <ClCompile Include="..\..\src\view\display-inventory.cpp" />
     <ClCompile Include="..\..\src\view\display-map.cpp" />
     <ClCompile Include="..\..\src\view\display-self-info.cpp" />
+    <ClCompile Include="..\..\src\view\display-scores.cpp" />
     <ClCompile Include="..\..\src\window\display-sub-window-items.cpp" />
     <ClCompile Include="..\..\src\window\display-sub-window-spells.cpp" />
     <ClCompile Include="..\..\src\window\display-sub-windows.cpp" />
@@ -1665,6 +1666,7 @@
     <ClInclude Include="..\..\src\view\display-messages.h" />
     <ClInclude Include="..\..\src\view\display-monster-status.h" />
     <ClInclude Include="..\..\src\view\display-self-info.h" />
+    <ClInclude Include="..\..\src\view\display-scores.h" />
     <ClInclude Include="..\..\src\window\display-sub-window-items.h" />
     <ClInclude Include="..\..\src\window\display-sub-window-spells.h" />
     <ClInclude Include="..\..\src\window\display-sub-windows.h" />
