+++ src/Makefile.am
@@ -881,6 +881,7 @@ hengband_SOURCES = \
 	view/display-player-misc-info.cpp view/display-player-misc-info.h \
 	view/display-player-middle.cpp view/display-player-middle.h \
 	view/display-player.cpp view/display-player.h \
+	view/display-scores.cpp view/display-scores.h \
 	view/display-self-info.cpp view/display-self-info.h \
 	view/display-store.cpp view/display-store.h \
 	view/display-util.cpp view/display-util.h \
