+++ src/view/display-scores.h
@@ -1,0 +1,5 @@
+﻿#pragma once
+
+struct high_score;
+void display_scores_aux(int from, int to, int note, high_score *score);
+void display_scores(int from, int to);
